Move selectors from shared/notifications into ui/ directory
