+++ ui/app/selectors/selectors.js
@@ -24,6 +24,7 @@ import {
 } from '../helpers/utils/conversions.util';
 
 import { TEMPLATED_CONFIRMATION_MESSAGE_TYPES } from '../pages/confirmation/templates';
+import { getSwapsFeatureLiveness } from '../ducks/swaps/swaps';
 
 import { getNativeCurrency } from './send';
 
@@ -497,4 +498,26 @@ export function getNextSuggestedNonce(state) {
 
 export function getShowWhatsNewPopup(state) {
   return state.appState.showWhatsNewPopup;
+}
+
+function getNotificationFilters(state) {
+  const currentNetworkIsMainnet = getIsMainnet(state);
+  const swapsIsEnabled = getSwapsFeatureLiveness(state);
+
+  return {
+    1: !currentNetworkIsMainnet || !swapsIsEnabled,
+  };
+}
+
+export function getSortedNotificationsToShow(state) {
+  const notifications = Object.values(state.metamask.notifications) || [];
+  const notificationFilters = getNotificationFilters(state);
+  const notificationsToShow = notifications.filter(
+    (notification) =>
+      !notification.isShown && !notificationFilters[notification.id],
+  );
+  const notificationsSortedByDate = notificationsToShow.sort(
+    (a, b) => new Date(b.date) - new Date(a.date),
+  );
+  return notificationsSortedByDate;
 }
